Refactored cli-commands module structure
