+++ src/cli-command/cli-command.interface.ts
@@ -1,4 +1,6 @@
-export interface CliCommandInterface {
+interface CliCommandInterface {
     readonly name: string;
     execute(...parameters: string[]): void;
-  }
+}
+
+export default CliCommandInterface;
