+++ src/cli-command/command-names.enum.ts
@@ -1,5 +1,7 @@
-export enum CommandNames {
+enum CommandNames {
   help = '--help',
   version = '--version',
   import = '--import',
 }
+
+export default CommandNames;
