+++ src/cli-command/version-command/version-command.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import { readFileSync } from 'fs';
-import { CliCommandInterface } from './cli-command.interface.js';
-import { CommandNames } from './command-names.enum.js';
+import CliCommandInterface from '../cli-command.interface.js';
+import CommandNames from '../command-names.enum.js';
 
 export default class VersionCommand implements CliCommandInterface {
   public readonly name = CommandNames.version;
