+++ src/cli.ts
@@ -1,9 +1,7 @@
 #!/usr/bin/env node
 
-import VersionCommand from './cli-command/version-command.js';
-import HelpCommand from './cli-command/help-command.js';
-import ImportCommand from './cli-command/import-command.js';
 import CLIApplication from './app/cli-application.js';
+import { HelpCommand, ImportCommand, VersionCommand } from './cli-command/index.js';
 
 const cli = new CLIApplication();
 cli.registerCommands([
